+++ halo/functions/object_unit.cpp
@@ -1,6 +1,6 @@
 #include "object_unit.hpp"
 #include "../../hooker/hooker.hpp"
-#include "../memory/object_unit.hpp"
+#include "../memory/gamestate/object/object_unit.hpp"
 
 // Avoid calling these if you aren't sure about what you're doing.
 // A LOT of these functions don't have any checks to see if you gave them valid input.
