+++ halo/functions/object_unit.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "../memory/object_unit.hpp"
+#include "../memory/gamestate/object/object_unit.hpp"
 
 __attribute__((regparm(3)))
 void unit_speak(MemRef object, int16_t arg2, UnitSpeech* speech_data);
