+++ halo/memory/gamestate/effect.cpp
@@ -1,23 +1,5 @@
+#include "effect.hpp"
 #include "table.hpp"
-
-#include "../../hooker/hooker.hpp"
-
-Signature(true, sig_object_table,
-    {-1, -1, -1, -1, 0x8B, 0x51, 0x34, 0x25, 0xFF, 0xFF, 0x00, 0x00, 0x8D, 0x04, 0x40, 0x8B, 0x44, 0x82, 0x08, 0x8B, 0x08});
-
-// Important, used for all effect related stuff
-Signature(true, sig_weather_particle_table,
-    {-1, -1, -1, -1, 0x89, 0x2D, -1, -1, -1, -1, 0x66, 0x89, 0x3D, -1, -1, -1, -1, 0x88, 0x5E, 0x24});
-
-Table** effect_ptrs_lst;
-
-Table** effect_ptrs(){
-    static uintptr_t sig_addr1 = sig_weather_particle_table.address();
-    if (!effect_ptrs_lst){
-        effect_ptrs_lst = reinterpret_cast<Table**>(sig_addr1);
-    };
-    return effect_ptrs_lst;
-}
 
 Table* weather_particles_table(){
     static Table** effect_ptrs_loc = effect_ptrs();
@@ -57,15 +39,4 @@ Table* contrail_points_table(){
 Table* contrail_table(){
     static Table** effect_ptrs_loc = effect_ptrs();
     return effect_ptrs_loc[7];
-}
-
-Table* device_groups_table(){
-    static Table** effect_ptrs_loc = effect_ptrs();
-    return effect_ptrs_loc[8];
-}
-
-ObjectTable* object_table(){
-    static ObjectTable** object_table_ptr =
-        reinterpret_cast<ObjectTable**>(sig_object_table.address());
-    return object_table_ptr[0];
 }
