+++ halo/memory/gamestate/effect.hpp
@@ -1,0 +1,11 @@
+#pragma once
+#include "table.hpp"
+
+Table* weather_particles_table();
+Table* particle_table();
+Table* particle_system_particles_table();
+Table* effect_table();
+Table* effect_locations_table();
+Table* decals_table();
+Table* contrail_points_table();
+Table* contrail_table();
