+++ halo/memory/gamestate/object/object_item.cpp
