+++ halo/memory/gamestate/object/object_item.hpp
@@ -1,5 +1,4 @@
 #pragma once
-#include "types.hpp"
 #include "object.hpp"
 
 #pragma pack(push, 1)
