+++ halo/memory/gamestate/player.cpp
@@ -1,0 +1,1 @@
+#include "player.hpp"
