+++ halo/memory/gamestate/terminal.cpp
@@ -1,0 +1,1 @@
+#include "terminal.hpp"
