--- halo/memory/object.cpp
@@ -1,1 +1,0 @@
-#include "object.hpp"
