--- halo/memory/object_device.cpp
@@ -1,1 +1,0 @@
-#include "object_device.hpp"
